Delivery/20251030/requirements.txt
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
black==25.9.0
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
entrypoints==0.4
Faker==37.12.0
favicon==0.7.0
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.11
langsmith==0.4.38
loguru==0.7.3
lxml==6.0.2
Markdown==3.9
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.9.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
postgrest==2.22.2
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdf==6.1.3
pytesseract==0.3.13
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
realtime==2.22.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==3.3.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.0.0
snowflake-snowpark-python==1.41.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
st-annotated-text==4.0.2
st-theme==1.2.3
storage3==2.22.2
streamlit==1.50.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.3
torch==2.5.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0; python_version < "3.13"
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
yarl==1.22.0
zstandard==0.25.0
